Config and service unit tests

+++ backend/models/card.py
@@ -2,7 +2,7 @@ from models import db
 
 class Card(db.Model):
     id = db.Column(
-        db.SmallInteger,
+        db.Integer,
         unique=True,
         nullable=False,
         primary_key=True,
