+++ backend/models/role.py
@@ -2,7 +2,7 @@ from models import db
 
 class Role(db.Model):
     id = db.Column(
-        db.SmallInteger,
+        db.Integer,
         unique=True,
         nullable=False,
         primary_key=True,
@@ -14,6 +14,6 @@ class Role(db.Model):
         nullable=False)
 
     level = db.Column(
-        db.SmallInteger,
+        db.Integer,
         unique=True,
         nullable=False)
